+++ CryptoLeo/CryptoLeo/Blockchain/Peer.swift
@@ -7,11 +7,15 @@
 
 import Foundation
 
+/// Node of the blockchain, which is able to transfer cryptocurrency or mine blocks and be rewarded.
+///
+/// - **Name**: Person's name.
+/// - **Public Key**: Personal public key, used for verifying transaction's signatures.
 struct Peer {
     
     /// Peer's name.
     let name: String
     
-    /// Peer's public key.
+    /// Peer's public key, used for verifying transaction's signatures.
     let publicKey: Data
 }
