+++ CryptoLeo/CryptoLeo/Blockchain/Reward.swift
@@ -8,6 +8,12 @@
 import Foundation
 import CryptoKit
 
+/// Reward given to a miner (`Peer`) for mining a block into and add it to the blockchain.
+///
+/// - **Miner**: Peer responsible for performing the needed computational work to mine a block.
+/// - **Amount**: Cryptocurrencies given as reward for mining a block.
+/// - **Timestamp**: Date and time of the reward.
+/// - **Message**: Description of the reward.
 struct Reward {
     
     /// Peer that's being rewarded with cryptocurrency.
@@ -19,6 +25,6 @@ struct Reward {
     /// When the transaction is being made.
     let timestamp: String = Timestamp.string()
     
-    /// Reward message.
+    /// Reward's description.
     let message: String
 }
